// src/app/interface/siteOwner/site.interface.ts
export interface SiteOwner {
   phone: string;
   phoneUrl: string;
   email: string;
   description: string;
   copyRight: string;
   abnNumber: string;
   abnUrl: string;
 };

 export interface YeonJoonKimInfo{
  completed: string;
  currentCompany: string;
  currentPosition: string;
  currentStudy: string;
  description: string;
  name: string;
  uid: string;
 };